chore: add lint, coverage, and fuzz gates and clear their findings
Adds golangci-lint, per-package coverage floors, and fuzz-corpus replay to
CI. The linter set is deliberately small — every rule catches a class of
defect this project has shipped or would tolerate silently — so that a
finding always means something rather than being style noise.

Clearing the first run found real problems. Four functions in the
conformance command were the superseded lifecycle-smoke path, dead since
the catalog runner replaced it, along with three unused helpers elsewhere;
all are removed. Response writes on both the JSON-RPC and MCP servers
discarded their encode errors, so a half-delivered frame to a broken pipe
was invisible; they now report on stderr, which is the only channel left
when the response channel itself has failed. Several errors stringified a
cause with %v while wrapping a sentinel with %w, which dropped the cause
from the error chain; they wrap both.

Where a linter was wrong, the code is left alone and the reason recorded:
lock releases in defers discard deliberately and now say so through a named
helper, QF1001 is suppressed because De Morgan's dual of the alias
character-class guard reads worse than the guard, and error strings naming
Claude are rephrased rather than lowercasing a product name.

Coverage floors start just under current: internal/cli 60% against 63.0%
measured, store 66% against 68.8%. The script is written for bash 3.2 so it
behaves the same on a maintainer's macOS shell as in CI.

+++ internal/cli/cmd_my.go
@@ -56,7 +56,7 @@ func cmdMy(args []string) {
 		}
 		leaseInfo := ""
 		if t.LeaseExpires > 0 {
-			remaining := int64(t.LeaseExpires) - int64(projectNow().UnixMilli())
+			remaining := int64(t.LeaseExpires) - projectNow().UnixMilli()
 			if remaining > 0 {
 				leaseInfo = fmt.Sprintf(" (lease: %s)", time.Duration(remaining)*time.Millisecond)
 			} else {
